wip: typescript migration. Should be mostly finished

+++ controllers/UserController.ts
@@ -1,14 +1,15 @@
 import bcrypt from 'bcrypt';
+import { Request, Response } from 'express';
 import mongoose, { ClientSession } from 'mongoose';
 import { Project } from '../database/models/project.js';
 import { User } from '../database/models/user.js';
-import { createToken } from '../utils/Generators.js';
-import { deleteProjectService } from './service/projectService.js';
-import { Request, Response } from 'express';
 import { LoginPayload, RegisterRequest } from '../types.js';
+import { createToken } from '../utils/Generators.js';
 import { createError, transact } from '../utils/Utils.js';
+import { deleteProjectService } from './service/projectService.js';
 
-const unableToDeleteUserString = "Unable to delete user. Please try again later."
+const unableToDeleteUserString =
+	'Unable to delete user. Please try again later.';
 
 export const createUser = async (req: Request, res: Response) => {
 	const { firstName, lastName, password, email } = req.body as RegisterRequest;
@@ -25,22 +26,21 @@ export const createUser = async (req: Request, res: Response) => {
 		email,
 	});
 
-
-	let token: string|Error = ""
+	let token: string | Error = '';
 	try {
 		await transact(mongoose, async (session: ClientSession) => {
-			await newUser.save({session});
+			await newUser.save({ session });
 			delete newUser.password;
-	
+
 			token = createToken(newUser);
 			if (token instanceof Error) {
-				throw new Error("Cannot create token when creating user")
+				throw new Error('Cannot create token when creating user');
 			}
 		});
 	} catch (error) {
-		return res.status(400).json(createError("Error registering"));
+		return res.status(400).json(createError('Error registering'));
 	}
-	const returnObject:LoginPayload = { token, user:newUser }
+	const returnObject: LoginPayload = { token, user: newUser };
 	res.status(200).send(returnObject);
 };
 
@@ -64,27 +64,25 @@ export const registerOAuthUser = async (req: Request, res: Response) => {
 
 export const deleteUser = async (req: Request, res: Response) => {
 	const userId = req.auth.id;
-	const session = await mongoose.startSession();
-	session.startTransaction();
 	try {
-		// Find all projects that have this user as their leader
-		const projects = await Project.find({ leader: userId }, null, { session });
-
-		// Iterate through each project
-		for (let project of projects) {
-			const deletedProject = await deleteProjectService(project._id, session);
-		}
-
-		// Finally delete the user
-		const deletedUser = await User.deleteOne({ _id: userId }, { session });
-
-		await session.commitTransaction();
-		session.endSession();
-
+		var deletedUser;
+		await transact(mongoose, async (session: ClientSession) => {
+			const projects = await Project.find({ leader: userId }, null, {
+				session,
+			});
+			if (!(projects.length > 0)) {
+				throw new Error('Cannot find project with leader');
+			}
+			for (let project of projects) {
+				await deleteProjectService(project._id, session);
+			}
+			deletedUser = await User.deleteOne({ _id: userId }, { session });
+			if (!deletedUser.acknowledged) {
+				throw new Error('Delete project service failure');
+			}
+		});
 		return res.status(200).send(deletedUser);
 	} catch (error) {
-		await session.abortTransaction();
-		session.endSession();
 		return res.status(500).json(createError(unableToDeleteUserString));
 	}
 };
